db: add new restore() method.
This will restore the database from an existing root.  This should work
whether or not the database actually exists or not.

It scans through the root directory for subdirectories.  It assumes the
subdirectories are named by docid, and those ids are used to recreate
documents.  Empty directories are ignored.  'bibtex' files in the directories
will be parsed as bibtex.  'tags' files will be used to set tags.  Any pdf
files will be added.