cli/nci: limit stdout codec manipulation to subcommands that really need it
Manipulating sys.stdout to support utf8 output was apparently breaking
readline, which is used in the 'add' command.  We limit this setting
to only the commands that need it, ie. the ones that are sending
output to stdout.  This has the benefit that we don't need to
remanipulate it when loading the nci interface.