db: change default OP to AND
This changes the search behavior by changing the default way in which
search terms are joined.  They were being joined with an OR, which
produces more search results when more terms are specified.  The default
now is to join terms with an AND, which reduces the results to documents
that include all terms.  I believe this is more intuitive.

The OR behavior can still be accessed by specifying it explicitly, e.g.:

search term1 OR term2