doc: much improved term iterator
This improved function is now a generator.  When supplied with a prefix it
iterators over just the prefixed terms.  It now takes advantage of term
adjacency by using the xapian term iterator skip_to() method to skip
directly to the the first prefixed term, and then breaking out of the loop
on the first non-matching term prefix.

This speeds up term searching *considerably*.  Search output is now more
than 4 times faster.